+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add function `dicetest::dice::shuffle_slice`. This allows to shuffle a slice in place.
 
 ## [0.2.1] - 2020-12-05
 ### Added